+++ src/mocks/friend.ts
